pyproject.toml
[project]
name = "wccc"
version = "0.1.0"
description = "WeChat API Client Tool"
authors = [
    { name = "andao" }
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.4",
    "psycopg2-binary>=2.9.10",
    "pymilvus[milvus-lite]>=2.6.3",
    "requests>=2.28.0",
    "sentence-transformers>=3.2.1",
    "torch>=2.5.1",
    "websocket-client>=1.5.0",
    "zai-sdk>=0.0.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["wechat", "ai", "bot", "db"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
